+++ internal/render/render.go
@@ -23,11 +23,14 @@ type Render struct {
 func New() *Render {
 	r := Render{
 		tmpls: template.New("render").Funcs(map[string]interface{}{
-			"Field":    fieldName,
-			"Receiver": receiverName,
-			"Param":    parameterName,
-			"Want":     wantName,
-			"Got":      gotName,
+			"Field":        fieldName,
+			"Receiver":     receiverName,
+			"Param":        parameterName,
+			"Want":         wantName,
+			"Got":          gotName,
+			"TypeArgs":     typeArguments,
+			"FieldType":    fieldType,
+			"ReceiverType": receiverType,
 		}),
 	}