+++ testdata/test039.go
@@ -1,0 +1,6 @@
+package testdata
+
+// Generic function with 'any' constraint
+func GenericAny[T any](val T) T {
+	return val
+}