+++ testdata/test040.go
@@ -1,0 +1,6 @@
+package testdata
+
+// Generic function with 'comparable' constraint
+func GenericComparable[T comparable](a, b T) bool {
+	return a == b
+}