+++ testdata/test043.go
@@ -1,0 +1,6 @@
+package testdata
+
+// Generic function with multiple type parameters
+func Pair[T, U any](first T, second U) (T, U) {
+	return first, second
+}